Update README.md
Low Energy Bluetooth Service Broadcasting
Prerequisite: Any IOT Development board like ESP32 or Arduino board with bluetooth on it.
TODO:
Write a code for the Dev. board that you choose to broadcast bluetooth services. (not a beacon) Services to broadcast:
1. Temperature Measurement
2. Humidity
Either interface the sensor with the development board or just write a function that mimic the sensor behaviour in code.